Move MainMenu script
